requirements.txt
nose
pyyaml
threadpool